fix(DeepSORT): fix code bug of numpy version.

+++ sample/DeepSORT/python/deep_sort/sort/detection.py
@@ -5,7 +5,7 @@ import numpy as np
 class Detection(object):
 
     def __init__(self, tlwh, cls_, confidence, feature):
-        self.tlwh = np.asarray(tlwh, dtype=np.float)
+        self.tlwh = np.asarray(tlwh, dtype=np.float32)
         self.cls_ = cls_
         self.confidence = float(confidence)
         self.feature = np.asarray(feature, dtype=np.float32)
